refactor(asciimode): Make ascii-mode a style opt
Before asciimode was a parameter flag. Now it is an option for
char-style.
It is called "Small" and setting the char-style to Small is equivalent
to turning asciimode off. Switching to other char-styles is equivalent
to turning asciimode on.
This simplifies logic and was an artifact from an early dev stage.